Move StoredDict class into new 'stored_dict' module
This commit is only moving code, no functional change
