+++ electrum/submarine_swaps.py
@@ -39,7 +39,7 @@ from .util import (
 from . import lnutil
 from .lnutil import hex_to_bytes, REDEEM_AFTER_DOUBLE_SPENT_DELAY, Keypair
 from .bolt11 import decode_bolt11_invoice
-from .json_db import StoredObject, stored_in
+from .stored_dict import StoredObject, stored_in
 from . import constants
 from .address_synchronizer import (TX_HEIGHT_LOCAL, TX_HEIGHT_FUTURE, TX_HEIGHT_UNCONFIRMED,
                                    TX_HEIGHT_UNCONF_PARENT)
